refactor: Extract type guards from validation module
Created src/lib/type-guards.ts module with 9 type guard functions:
- isArrayOf<T>()
- isObject()
- isString()
- hasProperty<K>()
- isNumber()
- isBoolean()
- isClarifierQuestion()
- isTask()
- isIdeaAnalysis()

Updated src/lib/validation.ts:
- Removed all type guard functions (130 lines removed)
- Added re-export of type guards for backward compatibility
- File size reduced from 338 to 208 lines (23% reduction)

Updated agent files:
- src/lib/agents/breakdown-engine/TaskDecomposer.ts
- src/lib/agents/breakdown-engine/IdeaAnalyzer.ts

SOLID Principles Applied:
- Single Responsibility: Type guards separated from validation logic
- Interface Segregation: Import only what you need
- Dependency Inversion: Dependencies flow correctly

Documentation:
- Updated docs/task.md with Task 5 completion

Testing:
- Build passes
- Type-check passes
- Validation tests pass (98/98)
- Breakdown engine tests pass (13/13)
- Zero breaking changes to production code
